fix(stackflow): move useStack to coordinator to prevent per-activity re-renders
useStack() in each activity's useHideEffect caused all activities to
re-render on any stack change, producing flicker during pop transitions.
Now useStack() is called once in GlobalInteraction via a centralized
useHideEffectCoordinator that directly manipulates DOM refs.
