+++ packages/stackflow/src/primitive/GlobalInteraction/GlobalInteraction.tsx
@@ -2,6 +2,7 @@ import { composeRefs } from "@radix-ui/react-compose-refs";
 import { mergeProps } from "@seed-design/dom-utils";
 import { Primitive, type PrimitiveProps } from "@seed-design/react-primitive";
 import { forwardRef } from "react";
+import { useHideEffectCoordinator } from "../../hooks/useHideEffectCoordinator";
 import { useGlobalInteraction } from "./useGlobalInteraction";
 import { GlobalInteractionProvider } from "./useGlobalInteractionContext";
 
@@ -12,6 +13,7 @@ export interface GlobalInteractionProps
 export const GlobalInteraction = forwardRef<HTMLDivElement, GlobalInteractionProps>(
   (props, ref) => {
     const api = useGlobalInteraction();
+    useHideEffectCoordinator();
     return (
       <GlobalInteractionProvider value={api}>
         <Primitive.div
